requirements.txt
flask  
requests  
google-generativeai  
sentence-transformers  
beautifulsoup4  
